Move main.rs to lib.rs
